Add a very rough draft of the add regimen screen
Still a lot of work to go. It's not very athentically pleasing.
Just hijacked the "profile" screen link to get a way to it. No state
is actually being saved, just a slew of UI elements.
Additionally, we might need to discuss how to structure the data.
I think it makes sense to have a starting place. But I'm not sure if
thats a datetime or just time. And an interval. But the interval could
be complex too, so for now just created a thin dropdown with a few choices.

ISSUES: issue-1